Strip unc prefixes
