+++ cli/npm/resolvers/local.rs
@@ -93,7 +93,9 @@ impl LocalNpmPackageResolver {
     match self.maybe_resolve_folder_for_specifier(specifier) {
       // Canonicalize the path so it's not pointing to the symlinked directory
       // in `node_modules` directory of the referrer.
-      Some(path) => Ok(self.fs.canonicalize(&path)?),
+      Some(path) => {
+        Ok(deno_core::strip_unc_prefix(self.fs.canonicalize(&path)?))
+      }
       None => bail!("could not find npm package for '{}'", specifier),
     }
   }
