+++ core/lib.rs
@@ -17,6 +17,7 @@ mod ops;
 mod ops_builtin;
 mod ops_builtin_v8;
 mod ops_metrics;
+mod path;
 mod realm;
 mod resources;
 mod runtime;
@@ -101,6 +102,7 @@ pub use crate::ops_builtin::op_resources;
 pub use crate::ops_builtin::op_void_async;
 pub use crate::ops_builtin::op_void_sync;
 pub use crate::ops_metrics::OpsTracker;
+pub use crate::path::strip_unc_prefix;
 pub use crate::realm::JsRealm;
 pub use crate::resources::AsyncResult;
 pub use crate::resources::Resource;
